pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "atlaswm"
version = "0.1.0"
description = "Stable end-to-end JEPA world models via deterministic distribution matching."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "Darvy C." },
]
keywords = [
  "jepa",
  "world-model",
  "self-supervised-learning",
  "representation-learning",
  "spherical-designs",
  "cramer-wold",
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
  "torch>=2.0",
  "numpy>=1.24",
  "tqdm>=4.65",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
student-t = ["scipy>=1.10"]
dev = [
  "pytest>=7.4",
  "pytest-cov>=4.1",
  "ruff>=0.3",
  "scipy>=1.10",
]

[project.urls]
Homepage = "https://github.com/darvyc/atlaswm"
Repository = "https://github.com/darvyc/atlaswm"
Issues = "https://github.com/darvyc/atlaswm/issues"

[project.scripts]
atlaswm-train = "scripts.train:main"

[tool.setuptools]
packages = ["atlaswm", "atlaswm.planning"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM"]
ignore = ["E501"]  # allow long lines for math expressions
